requirements.txt
Django==2.2.24
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
pillow==8.0
